+++ src/components/FlawLabels/FlawLabelsTable.vue
@@ -39,6 +39,13 @@ const availableBuLabels = computed(() =>
   ),
 );
 
+const nonEditableLabelTypes: FlawLabelTypeEnum[] = [
+  FlawLabelTypeEnum.WORKFLOW,
+  FlawLabelTypeEnum.PRODUCT_FAMILY,
+  FlawLabelTypeEnum.ALIAS,
+];
+const isNonEditableLabelType = (type: FlawLabelTypeEnum) => nonEditableLabelTypes.includes(type);
+
 const isExpandedDefault = labelsFromProps.value.some(label => label.contributor || label.state === 'NEW');
 const [isExpanded, toggleExpanded] = useToggle(isExpandedDefault);
 
@@ -112,17 +119,17 @@ function handleUndoDelete(label: ZodFlawLabelType) {
             <td
               :class="{
                 'fw-bold': label.state === 'REQ',
-                'text-decoration-line-through': !label.relevant,
+                'text-decoration-line-through': label.relevant === false,
               }"
-              :title="!label.relevant ? 'Associated affect was removed' : undefined"
+              :title="label.relevant === false ? 'Associated affect was removed' : undefined"
             >{{ label.name }}</td>
             <td>{{ label.contributor }}</td>
             <td>
               <div class="actions">
                 <template v-if="!isDeletedLabel(label)">
                   <span
-                    v-if="label.type === FlawLabelTypeEnum.WORKFLOW && !isNewLabel(label)"
-                    title="Workflow labels are not editable"
+                    v-if="isNonEditableLabelType(label.type)"
+                    :title="`${label.type} labels are not editable`"
                   >
                     <button
                       type="button"