+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 # OSIM Changelog
+## [Unreleased]
+### Fixed
+* Make workflow, product_family and alias labels non-editable (`OSIDB-5385`)
+* Fix strikethrough incorrectly applied to labels without relevant field (`OSIDB-5375`)
+
 ## [2026.8.0-patch-missing-labels]
 ### Fixed
 * New labels can now be displayed on flaw page and queried on advanced search